improve ExprIntLiteral.autoType

+++ src/main/resources-test/mine.c
@@ -230,23 +230,23 @@ void main() {
 		}
 
 		// cursor up
-		if (chr == 0xE048) {
+		if (chr == (i16)0xE048) {
 			curr_r = (curr_r + height - 1) % height;
 		}
 		// cursor down
-		else if (chr == 0xE050) {
+		else if (chr == (i16)0xE050) {
 			curr_r = (curr_r + 1) % height;
 		}
 		// cursor left
-		else if (chr == 0xE04B) {
+		else if (chr == (i16)0xE04B) {
 			curr_c = (curr_c + width - 1) % width;
 		}
 		// cursor left
-		else if (chr == 0xE04B) {
+		else if (chr == (i16)0xE04B) {
 			curr_c = (curr_c + width - 1) % width;
 		}
 		// cursor right
-		else if (chr == 0xE04D) {
+		else if (chr == (i16)0xE04D) {
 			curr_c = (curr_c + 1) % width;
 		}
 		// space = flag
